% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaintexp.r
\name{all_human_genes}
\alias{all_human_genes}
\title{Get human gene symbols from biomaRt}
\usage{
all_human_genes()
}
\value{
A data frame with information about all human genes taken from biomaRt: Chomosome, start and end position, strand and the HGNC gene symbol.
}
\description{
Get human gene symbols from biomaRt
}
