% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaintexp.r
\name{CiberAMP.classifier}
\alias{CiberAMP.classifier}
\title{Perform joint expression and copy-number variation analyses on a set of genes using a set of tumor samples.}
\usage{
CiberAMP.classifier(res1 = NULL, res3 = NULL, width.window = 1e+06)
}
\arguments{
\item{res1}{The first resulting data frame from CiberAMP outcomes}

\item{res3}{The third resulting data frame from CiberAMP outcomes}

\item{width.window}{Number of base pairs to be analyzed in order to define genomic clusters}
}
\value{
List containing four data frames:
1 - Top candidates: SCN-associated DEGs with no co-amplification/deletion with a known oncogene and OUTSIDE a genomic cluster
2 - Second best candidates: SCN-associated DEGs with no co-amplification/deletion with a known oncogene and INSIDE a genomic cluster
3 - Third best candidates: SCN-associated DEGs co-amplified/deleted with a known oncogene and OUTSIDE a genomic cluster. This may contain COSMIC CGC oncogenes.
4 - Fourth best candidates: SCN-associated DEGs co-amplified/deleted with a known oncogene and INSIDE a genomic cluster. This contains COSMIC CGC oncogenes.
}
\description{
Perform joint expression and copy-number variation analyses on a set of genes using a set of tumor samples.
}
