% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaintexp.r, R/load_data.R
\docType{data}
\name{ciberAMP}
\alias{ciberAMP}
\title{Perform joint expression and copy-number variation analyses on a set of genes using a set of tumor samples.}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
ciberAMP(
  genes = c(),
  cohorts = c(),
  writePath = NULL,
  pat.percentage = 0,
  pp.cor.cut = 0.6,
  norm.method = "geneLength",
  filt.method = "quantile",
  filt.qnt.cut = 0.25,
  filt.var.func = "IQR",
  filt.var.cutoff = 0.75,
  filt.eta = 0.05,
  filt.FDR.DEA = 0.05,
  filt.FC = 0.58,
  cna.thr = "Deep",
  exp.mat = NULL,
  cna.mat = NULL
)

data(ciberAMP)
}
\arguments{
\item{genes}{Vector of approved queried genes symbols to be analyzed.}

\item{cohorts}{Vector of TCGA cohorts IDs. See function \code{all_tumors} for a list of available tumors. If is left empty, CiberAMP will run on every cohort.}

\item{writePath}{Path where results are stored. Defaults to the current folder.}

\item{pat.percentage}{Minimum percentage of patients per group.}

\item{pp.cor.cut}{Threshold to filter samples by AICC. Passed to `TCGAanalyze_Preprocessing`.}

\item{norm.method}{Method of normalization, such as `gcContent` or `geneLength` (default).
Passed to `TCGAanalyze_Normalization`.}

\item{filt.method}{Method of filtering, such as `quantile` (default), `varFilter`, `filter1`, `filter2`.}

\item{filt.qnt.cut}{Threshold selected as mean for filtering. Defaults to 0.25.}

\item{filt.var.func}{Filtering function. Defaults to `IQR`. See `genefilter` documentation for available methods.}

\item{filt.var.cutoff}{Threshold for `filt.var.funct`.}

\item{filt.eta}{Parameter for `filter1`. Defaults to 0.05.}

\item{filt.FDR.DEA}{Threshold to filter differentially expressed genes according their adjusted p-value.
Passed to `TCGAanalyze_DEA`.}

\item{filt.FC}{Minimum log2(FC) value to considered a gene as differentially expressed. Defaults to 1.}

\item{cna.thr}{Threshold level for copy-number variation analysis. Can be `Deep`, `Shallow` or `Both`.
`Deep`, to consider homozygous deletions and high-level broad amplifications.
`Shallow`, to consider hemyzygous deletions and low-level focal amplifications.
`Both`, to consider deep and shallow amplifications/deletions in the same group: one group -> deep + shallow amp; second group -> deep + shallow del.}

\item{exp.mat}{Custom normalized RNAseq counts expression matrix of only tumors. Defaults to `NULL`.}

\item{cna.mat}{Custom copy-number analysis matrix of only tumors. Defaults to `NULL`.}
}
\value{
List containing three data frames:
1 - SCNA-mRNA correlations for queried genes
2 - SCNA-mRNA correlations for Cancer Census genes (COSMIC) | Known oncodriver/TSGs.
3 - SCNA overlapping between queried and CGC genes.
}
\description{
Perform joint expression and copy-number variation analyses on a set of genes using a set of tumor samples.

CiberAMP LUSC/HNSC example data (genes)
}
\keyword{datasets}
