% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaintexp.r
\name{CNAintEXP}
\alias{CNAintEXP}
\title{Perform joint expression and copy-number variation analyses on a set of genes using a set of tumor samples.}
\usage{
CNAintEXP(
  genes = c(),
  cohorts = c(),
  writePath = NULL,
  pat.percentage = 10,
  pp.cor.cut = 0.6,
  norm.method = "geneLength",
  filt.method = "quantile",
  filt.qnt.cut = 0.25,
  filt.var.func = "IQR",
  filt.var.cutoff = 0.75,
  filt.eta = 0.05,
  filt.FDR.DEA = 0.01,
  filt.FC = 1,
  normality.thr = 0.05,
  var.thr = 0.05,
  p.val.thr = 1,
  cna.thr = "Deep",
  exp.mat = NULL,
  cna.mat = NULL
)
}
\arguments{
\item{genes}{List of official gene symbols for analysis. See function \code{all_genes} for a list of available genes.}

\item{cohorts}{List of tumors to study. See function \code{all_tumors} for a list of available tumors.
Defaults to all tumors with either an empty list or a list containing `ALL`.}

\item{writePath}{Path where results are stored. Defaults to the current folder.}

\item{pat.percentage}{Minimum percentage of patients per group.}

\item{pp.cor.cut}{Threshold to filter samples by AICC. Passed to `TCGAanalyze_Preprocessing`.}

\item{norm.method}{Method of normalization, such as `gcContent` or `geneLength` (default).
Passed to `TCGAanalyze_Normalization`.}

\item{filt.method}{Method of filtering, such as `quantile` (default), `varFilter`, `filter1`, `filter2`.}

\item{filt.qnt.cut}{Threshold selected as mean for filtering. Defaults to 0.25.}

\item{filt.var.func}{Filtering function. Defaults to `IQR`. See `genefilter` documentation for available methods.}

\item{filt.var.cutoff}{Threshold for `filt.var.funct`.}

\item{filt.eta}{Parameter for `filter1`. Defaults to 0.05.}

\item{filt.FDR.DEA}{Threshold to filter differentially expressed genes according their corrected p-value.
Passed to `TCGAanalyze_DEA`.}

\item{filt.FC}{Parameter for `filter2`. Defaults to 1.}

\item{normality.thr}{Threshold for normality tests. Defaults to 0.05.}

\item{var.thr}{Threshold for tests of equal variance. Defaults to 0.05.}

\item{p.val.thr}{Threshold for reported p values. Defaults to 1 (report all).}

\item{cna.thr}{Threshold level for copy-number variation analysis. Can be `Deep`, `Shallow` or `Both`.
`Deep`, consider likely homozygous deletions for loss and high-level amplifications (high copy-number, focal) for gain.
`Shallow`, consider likely heterozygous deletions for loss and low-level amplifications (lower copy-number, broad) for gain.
`Both`, consider all loss and gain events.}

\item{exp.mat}{Custom expression matrix, which will not be filtered. Defaults to `NULL`,
which means that the method considers TCGA expression matrices belonging to `cohorts`.}

\item{cna.mat}{Custom copy-number analysis matrix, which will not be filtered. Defaults to `NULL`,
which means that the method considers TCGA copy-number analysis matrices belonging to `cohorts`.}

\item{dea.method}{Differential expression analysis method. Either `glmLRT` or `exactTest`.
`glmLRT`: Fit a negative binomial generalized log-linear model to the read counts for each gene.
`exactTest`: Compute genewise exact tests for differences in the means between two groups of negative-binomially
distributed counts.}
}
\value{
List containing three data frames:
1 - SCNA-mRNA correlations for queried genes
2 - SCNA-mRNA correlations for Cancer Census genes (COSMIC) | Known oncodriver/TS genes.
3 - SCNA overlapping between queried and CGC genes.
}
\description{
Perform joint expression and copy-number variation analyses on a set of genes using a set of tumor samples.
}
