% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaintexp.r
\name{ggplot.CiberAMP}
\alias{ggplot.CiberAMP}
\title{Plot CiberAMP results with ggplot}
\usage{
ggplot.CiberAMP(output)
}
\arguments{
\item{output}{List of results from CNAintEXP function.}
}
\value{
ggplot2 graph where Y axis = mRNA diff. expression between SCN-altered vs diploit tumors and X axis = mRNA diff. expression between Tumor and Normal tissue
}
\description{
Plot CiberAMP results with ggplot
}
